Modular Thrusters
The enemy ship's thrusters actually apply forces based off of where they are, and that will change if you move the thrusters around in the editor. Also I fixed a bug that caused the player to spawn in random places around the ship. And I made the camera zoom when you respawn a bit better.