Added a background
There is now a picture of Saturn in the background. The colors look a bit washed out because I make it darker to increase the contrast with the foreground. In the future I may want to replace it with a 3D model. I also fixed some issues with the minimap/radar thing showing shadows, but it still somewhat shows lighting, which I don't know what to do about.